+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 Unreleased
 
 - Replaced use of "operand" with the more accurate "identifier"
+- Updates seprating characters to match boon spec